pyproject.toml
[tool.poetry]
name = "transaction-reporting"
version = "0.1.0"
description = "Django backend for transaction reporting and aggregation."
authors = ["Sergii Denysiuk <sergii.denysiuk@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.2.8"
djangorestframework = "^3.16"
inflection = "^0.5.1"  # for DRF get_schema_view()
uritemplate = "^4.2.0"  # for DRF get_schema_view()
pyyaml = "^6.0.3"  # for DRF get_schema_view()
psycopg = {extras = ["binary"], version = "3.3.0"}
django-environ = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-django = "^4.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "transaction_reporting.settings"
python_files = [
  "tests.py",
  "test_*.py",
  "*_tests.py",
]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
line-length = 88  # Match Black line length for consistency
fix = true  # Allow Ruff to automatically apply safe fixes

[tool.ruff.lint]
# Enabled rule categories:
select = [
  "E",   # Pycodestyle errors (e.g., indentation, whitespace, line breaks)
  "F",   # Pyflakes errors (e.g., unused imports, undefined names)
  "I",   # Import sorting and organization (isort-compatible rules)
  "B",   # Bugbear: likely bugs and design issues
  "UP",  # Python upgrade: modernization and use of newer syntax
  "SIM", # Code simplifications and redundant patterns
  "PLE", # Pylint errors (high-severity issues)
  "C90", # McCabe complexity checks (e.g., overly complex functions)
]

# Ignored rules:
ignore = [
  "E203", # Whitespace before ':' (Black-compatible; Black formats like this on purpose)
  "E501", # Line too long (Black manages line length; rely on it instead of Ruff here)
]

[tool.mypy]
python_version = "3.13"
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
